+++ __tests__/entities/scene.entity.spec.ts
@@ -1,21 +1,20 @@
-import {
-  LoadedScene,
-  SessionControlResponseEvent,
-} from '@proto/ai/inworld/packets/packets_pb';
+import { LoadedScene } from '@proto/ai/inworld/packets/packets_pb';
 
 import { Character } from '../../src/entities/character.entity';
 import { Scene } from '../../src/entities/scene.entity';
 import { createAgent, createCharacter, SCENE } from '../helpers';
 
 let characters: Array<Character> = [];
+let name: string;
 let scene: Scene;
 let json: string;
 
 beforeEach(() => {
   jest.clearAllMocks();
 
+  name = SCENE;
   characters = [createCharacter(), createCharacter()];
-  scene = new Scene({ name: SCENE, characters });
+  scene = new Scene({ name, characters });
   json = JSON.stringify(scene);
 });
 
@@ -37,9 +36,7 @@ test('should deserialize', () => {
 test('should convert proto to scene', () => {
   const agents = [createAgent(), createAgent(false)];
 
-  const proto = new SessionControlResponseEvent().setLoadedScene(
-    new LoadedScene().setAgentsList(agents),
-  );
+  const proto = new LoadedScene().setAgentsList(agents);
   const scene = Scene.fromProto(SCENE, proto);
 
   expect(scene.characters[0].id).toEqual(agents[0].getAgentId());