+++ src/services/connection.service.ts
@@ -1,6 +1,10 @@
 import { ClientDuplexStream, ServiceError } from '@grpc/grpc-js';
 import { ClientRequest } from '@proto/ai/inworld/engine/world-engine_pb';
-import { InworldPacket as ProtoPacket } from '@proto/ai/inworld/packets/packets_pb';
+import {
+  InworldPacket as ProtoPacket,
+  LoadedCharacters,
+  LoadedScene,
+} from '@proto/ai/inworld/packets/packets_pb';
 
 import {
   ApiKey,
@@ -50,8 +54,8 @@ export class ConnectionService<
 > {
   private state: ConnectionState = ConnectionState.INACTIVE;
 
+  private nextSceneName: string | undefined;
   private scene: Scene;
-  private sceneIsLoaded: boolean = false;
   private sessionToken: SessionToken;
   private stream: ClientDuplexStream<ProtoPacket, ProtoPacket>;
   private connectionProps: ConnectionProps<InworldPacketT>;
@@ -85,6 +89,14 @@ export class ConnectionService<
     this.onError = this.connectionProps.onError;
 
     this.onMessage = async (packet: ProtoPacket) => {
+      const sessionControlResponse = packet.getSessionControlResponse();
+
+      if (sessionControlResponse?.hasLoadedScene()) {
+        this.setSceneFromProtoEvent(sessionControlResponse.getLoadedScene());
+      } else if (sessionControlResponse?.hasLoadedCharacters()) {
+        this.addCharactersToScene(sessionControlResponse.getLoadedCharacters());
+      }
+
       const inworldPacket = this.convertPacketFromProto(packet);
 
       this.connectionProps.onMessage?.(inworldPacket);
@@ -119,6 +131,10 @@ export class ConnectionService<
 
   getSceneName() {
     return this.scene.name;
+  }
+
+  setNextSceneName(name?: string) {
+    this.nextSceneName = name;
   }
 
   async generateSessionToken() {
@@ -180,9 +196,7 @@ export class ConnectionService<
   }
 
   async getCharactersList() {
-    if (!this.sceneIsLoaded) {
-      await this.open();
-    }
+    await this.open();
 
     return this.scene.characters;
   }
@@ -203,9 +217,10 @@ export class ConnectionService<
     try {
       await this.loadToken();
 
-      const { client, name, sessionContinuation, user } = this.connectionProps;
+      const { client, sessionContinuation, user } = this.connectionProps;
 
-      const [stream, scene] = await this.engineService.openSession({
+      const name = this.getSceneName();
+      const [stream, sessionProto] = await this.engineService.openSession({
         client,
         name,
         sessionContinuation,
@@ -218,20 +233,10 @@ export class ConnectionService<
         ...(this.onMessage && { onMessage: this.onMessage }),
       });
 
+      this.setSceneFromProtoEvent(sessionProto);
       this.stream = stream;
-      this.scene = scene;
-      this.sceneIsLoaded = true;
-
-      if (
-        !this.getEventFactory().getCurrentCharacter() &&
-        this.scene.characters.length
-      ) {
-        this.getEventFactory().setCurrentCharacter(this.scene.characters[0]);
-      }
-
       this.state = ConnectionState.ACTIVE;
       this.releaseQueue();
-
       this.scheduleDisconnect();
     } catch (err) {
       this.onError(err);
@@ -409,5 +414,37 @@ export class ConnectionService<
     return this.connectionProps.extension?.convertPacketFromProto
       ? this.connectionProps.extension.convertPacketFromProto(packet)
       : (InworldPacket.fromProto(packet) as InworldPacketT);
+  }
+
+  private ensureCurrentCharacter() {
+    const factory = this.getEventFactory();
+    const currentCharacter = factory.getCurrentCharacter();
+    const sameCharacter = currentCharacter
+      ? this.scene.characters.find(
+          (c) => c.resourceName === currentCharacter?.resourceName,
+        )
+      : undefined;
+
+    factory.setCurrentCharacter(sameCharacter ?? this.scene.characters[0]);
+  }
+
+  private setSceneFromProtoEvent(proto: LoadedScene) {
+    const name = this.nextSceneName || this.getSceneName();
+
+    this.scene = Scene.fromProto(name, proto);
+
+    this.connectionProps.extension?.afterLoadScene?.(proto);
+    this.ensureCurrentCharacter();
+  }
+
+  private addCharactersToScene(proto: LoadedCharacters) {
+    const characters = proto.getAgentsList().map((c) => Character.fromProto(c));
+
+    this.scene = new Scene({
+      name: this.scene.name,
+      characters: [...new Set(this.scene.characters.concat(characters))],
+    });
+
+    this.ensureCurrentCharacter();
   }
 }